BugID:16959715: reduce codesize for LoRaWAN 4.4.2

+++ network/lorawan_4_4_2/lora/mac/region/RegionUS915-Hybrid.c
@@ -592,7 +592,7 @@ bool RegionUS915HybridAdrNext( AdrNextParams_t* adrNext, int8_t* drOut, int8_t* 
 
 void RegionUS915HybridComputeRxWindowParameters( int8_t datarate, uint8_t minRxSymbols, uint32_t rxError, RxConfigParams_t *rxConfigParams )
 {
-    double tSymbol = 0.0;
+    DECIMAL tSymbol = 0.0;
     uint32_t radioWakeUpTime;
 
     rxConfigParams->Datarate = datarate;
